Add proper long line wrapping.
When sending a message to IRC, the message is relayed to other
servers and other clients by the server and needs to fit within
the max line length for server messages (512). What this means
is that the actual content of the message needs to be truncated
based on the length of the user's nick, hostname, and the kind
of message being sent (PRIVMSG, NOTICE). The current code,
in addition to hardcoding the wrap length at the absolute
max line length (not taking any of this into consideration),
is rather ugly and hard to follow. I've abstracted the logic
into something easier to follow, correct, and reuseable. It'll
be easy to expand it to ACTION.

TODO: maintain colors and other formatting across wrapped
messages. Textual does this for PRIVMSGs, at least, so we can try
looking there for inspiration.