Fix TLS connections.
Root cause: the `self.opt.selfSigned` parameter wasn't being
assigned to the options passed to `tls.connect()`, meaning that
connections to most ircds would fail. This explains why
connections to freenode (who have a proper certificate) work
while connections to most other ircds (which use self-signed
certs) fail.

While here, write some tests that ensure connecting to a secure
server actually works. This involves a refactor of
tests/test-irc.js to support both plaintext and encrypted
connections. This also includes a self-signed certificate for
testing locally.

There are also a few small, minor fixes to the helpers to either
remove unnecessary code or clean up. MockIrcd now supports an
option to its constructor that determines whether the connection
should be secure or not.

Still to do: refactor all of the TLS handling in irc.js to be
cleaner and to remove the duplicate code.