Fix a bug which prevents a client from connecting if their nick == NICKLEN
To resolve nick conflicts we add on a monotonically increasing numeric suffix.
This is appended to the original desired nick. However, if the original nick
is equal to the NICKLEN, it is truncated before being checked for clashes. This
then means the new proposed nick fails, ad infinitum.

This is complicated because we don't know what the NICKLEN is initially because
it is only sent as a rpl_isupport AFTER you connect successfully, which
requires... a valid nick! To work around this, we remember what we tried to
change the nick to, and then compare it to the nick which is returned in
err_nicknameinuse. If they don't match, we know we've been truncated and can
assume that characters UP TO the nick length in the *ERROR* response is allowed.