Fix SASL on libera.chat
While deploying a bot to libera.chat we were unable to connect via SASL because the ACK was never recognized. The reason for that rested on this weird space.

This patch checks that a trimmed version of the second argument is equal to `sasl`. This is enough for our personal purposes.